Change the return type of the open-ils.qstore.prepare method.
Formerly, this method merely returned a string to be used as 
token for future references to the query.

Now it returns an object, with the following elements:

"token" -- the query token as returned formerly
"bind_variables" -- a (possibly empty) object containing a
list of bind variables and their various attributes, as
returned by the param_list method.

Rationale: an interactive query-building client will always
want a list of bind variables so that it can prompt the user
for values.  With this change to the return type, the client
can get such a list without having to make a separate call
to the param_list method.

M    Open-ILS/src/c-apps/oils_qstore.c


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16653 dcc99617-32d9-48b4-a31d-7c20da2025e4