Booking: begin forward-porting code from rel_1_6.
Booking (regrettably) was largely written directly against a 1.6 environment
instead of written for trunk and backported.  So now we have booking code in
rel_1_6 and rel_1_6_1 that works, but that needs to be cleanly merged with
trunk.  There has been a lot of drift, and this won't be easy.

Here is the first step (some of the easy stuff).  These files have been updated
wholesale with their contents from rel_1_6, since they don't affect anything
other than booking itself.

Just to be clear: this commit does not complete the booking foward-port.  The
booking module did not work in trunk before this commit, and it does not work
after this commit.  For the moment, booking only works in the rel_1_6 branch,
in the rel_1_6_1 branch, and in 1.6.1.* releases.  It still does not work in
trunk, and it will take a few more hairy commits to get things in sync.

Once that's finally done, any future Booking code can be written the Right Way
(in trunk) and *back*ported thence to whatever other branches as needed.



git-svn-id: svn://svn.open-ils.org/ILS/trunk@16827 dcc99617-32d9-48b4-a31d-7c20da2025e4