Streamline pcrud a bit.
If there is a permacrud entry for the relevant action, but the list of
permissions is empty, we grant access.

The change: grant access as soon as the permission list can be known
to be empty, instead of doing a lot of irrelevant stuff first
(including extra database fetches in some cases).

M    Open-ILS/src/c-apps/oils_sql.c


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16880 dcc99617-32d9-48b4-a31d-7c20da2025e4