Prevent MARC batch import from blindly trusting the user by checking LDR/06
With this commit, Vandelay will now check the leader of the incoming MARC
record to ensure that its MARC type matches the indicated import record
type. This is useful, for example, if you have a mixed set of bib & holdings
records that you want to import; you don't want to import the holdings records
as bib records.

This is a fairly strict implementation; we could relax this by turning the
check on its head and allowing the import as the user-indicated type as long
as the LDR/06 doesn't explicitly match another record type.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16914 dcc99617-32d9-48b4-a31d-7c20da2025e4