Switch to ingesting authority records automatically via database triggers
Taken largely from the existing approach for ingesting biblio records,
there is some duplication here that we could genericize. But for now,
we'll go with the working-but-separate approach to avoid introducing
badness into the relatively stable biblio ingest pipeline.

Also, move some of the existing authority-related functions out of
the metabib schema file and into 999.functions.global.sql.

Finally, the direct_ingest.pl script is no longer necessary for
ingest of authority records, either, so die in that case as well.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16970 dcc99617-32d9-48b4-a31d-7c20da2025e4