The old default skin is the new default skin
We need people willing to maintain the craftsman skin, as it is falling
behind "default" in functionality and goes for long periods in a broken state.

This change gives Evergreen the most functional and most likely to work skin
by default, which is important for making positive first impressions on
possible adopters of the project.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@17096 dcc99617-32d9-48b4-a31d-7c20da2025e4