Move to CSS-based styling of validated vs. unvalidated fields
This sets the stage for differentiating between validated fields that
match authority records, and validated fields with a controlling $0
subfield that match an authority record. This also gives sites a bit
of an easier entry point to customize their MARC editor - say, for example,
to support accessibility requirements if red vs. black doesn't satisfy
the needs of those who are colour-blind.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@17234 dcc99617-32d9-48b4-a31d-7c20da2025e4