Booking: finish the forward-port from rel_1_6
This /should/ complete the forward port of booking from the rel_1_6 branch,
which means that booking in trunk should work just how it does in the latest
1.6.1.* releases.

Most of the changes in this commit were to Circ/Circulate.pm, and cursory
tests don't indicate any problems in the circulation logic overall, but the
lay of the land there is quite different now in trunk than it was when Booking
was initially developed, so I'd be somewhat wary for a little while.

Going forward, trunk can accept improvements/bug fixes/etc for booking, and
those changes can be *back*ported to branches in the usual way.  Yay!


git-svn-id: svn://svn.open-ils.org/ILS/trunk@17297 dcc99617-32d9-48b4-a31d-7c20da2025e4