Add label_prefix and label_suffix columns to asset.copy_location
These columns will be used to (optionally) set label prefixes and suffixes
when printing spine labels for copies at a given copy location.

The copy location editor has been taught to be aware of these new columns,
but right now a NULL prefix or suffix will be turned into an empty string
once the copy location has been edited. That might not be a big deal, but
I'm a foolish hobgoblin and like consistency.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@17363 dcc99617-32d9-48b4-a31d-7c20da2025e4