Add a bib_source editor widget to the MARC editor
This widget should only be exposed when editing bibliographic MARC records.
The existing restriction of not being able to set the bib source when importing
via Z39.50 sadly remains in place; I believe the best place to tackle that will
be in the Z39.50 import interface itself.

Also TODO: i18n


git-svn-id: svn://svn.open-ils.org/ILS/trunk@17393 dcc99617-32d9-48b4-a31d-7c20da2025e4