use circ_lib param instead of direct ws_ou acccess
circulator->circ_lib, which defaults to requestor->ws_ou (but is
overrideable) should be used in in lieu of direct ws_ou accesss;

This is useful when the caller is not physically logged in to a
workstation (e.g. centralized SIP client) and need to specify the
physical location through other means.

This change has no effect unless circ_lib is overridden by the caller

git-svn-id: svn://svn.open-ils.org/ILS/trunk@17400 dcc99617-32d9-48b4-a31d-7c20da2025e4