Remove the previously deprecated CGI interfaces for configuring the Evergreen system
The CGI interfaces have not been maintained; the interfaces available through
the Admin -> Server Administration menu in the staff client are the recommended
method for setting up new libraries in the organizational hierarchy,
permissions, copy statuses, and circulation rules.

Note that an adjustment to eg_vhost.conf is recommended to point to the
offline.pl script, which is the only remaining CGI script in use; this
should avoid conflicting Apache definitions for the /cgi-bin/ alias.

Addresses https://bugs.launchpad.net/evergreen/+bug/634984


git-svn-id: svn://svn.open-ils.org/ILS/trunk@17574 dcc99617-32d9-48b4-a31d-7c20da2025e4