Overhaul ORDRSP processing based on new Business::EDI capabilities
Lots of error checking

Fetch updates for xpath_value and a B&T data hack

ORDRSP is forced only b/c the DB requires it.  We can't accurately
know the message type until the translator deals with it (or we
build our own fault-prone sniffer).  But for now all we expect from
vendors in ORDRSP, so we can force temporarily.

Also EDI example data and debugging tweaks

git-svn-id: svn://svn.open-ils.org/ILS/trunk@17671 dcc99617-32d9-48b4-a31d-7c20da2025e4